+++ tests/Models/Product.php
@@ -1,0 +1,16 @@
+<?php
+
+namespace ServiceTo\UsesDetail\Tests\Models;
+
+use Illuminate\Database\Eloquent\Model;
+
+class Product extends Model
+{
+    protected $table = 'products';
+    protected $guarded = [];
+
+    public function categories()
+    {
+        return $this->belongsToMany(Category::class, 'category_product', 'product_id', 'category_id');
+    }
+}